requirements.txt
branca==0.3.1
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
decorator==4.3.0
dill==0.2.8.2
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
folium==0.7.0
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gcloud==0.18.3
gcloud-requests==1.1.9
google-api-core==1.5.2
google-auth==1.6.0
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.7.0
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
google-cloud-storage==1.13.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.16.1
httplib2==0.12.0
httplib2shim==0.0.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.3
oauth2client==3.0.0
pandas==0.23.4
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.5
pytz==2018.6
requests==2.20.1
retry==0.9.2
rsa==3.4.2
selenium==3.141.0
six==1.11.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
slackweb==1.0.5